+++ src/Keylight/Keylight.cs
@@ -177,6 +177,10 @@ namespace Keylight {
         AppVersion   = _config.AppVersion,
         SdkVersion   = SdkInfo.Version,
         Platform     = _config.Platform ?? Device.Platform,
+        CpuCores     = Device.CpuCores,
+        OsVersion    = Device.OsVersionValue,
+        Arch         = Device.Arch,
+        Memory       = Device.Memory,
         SdkTrialDurationDays = _config.TrialDurationDays
       };
 
@@ -230,6 +234,10 @@ namespace Keylight {
         AppVersion = _config.AppVersion,
         SdkVersion = SdkInfo.Version,
         Platform   = _config.Platform ?? Device.Platform,
+        CpuCores   = Device.CpuCores,
+        OsVersion  = Device.OsVersionValue,
+        Arch       = Device.Arch,
+        Memory     = Device.Memory,
         SdkTrialDurationDays = _config.TrialDurationDays
       };