+++ unity/dev.keylight.sdk/Runtime/Core/Keylight.cs
@@ -176,7 +176,11 @@ namespace Keylight {
         InstanceName = Device.DefaultInstanceName,
         AppVersion   = _config.AppVersion,
         SdkVersion   = SdkInfo.Version,
-        Platform     = _config.Platform ?? Device.Platform
+        Platform     = _config.Platform ?? Device.Platform,
+        CpuCores     = Device.CpuCores,
+        OsVersion    = Device.OsVersionValue,
+        Arch         = Device.Arch,
+        Memory       = Device.Memory
       };
 
       ActivateResponse resp;
@@ -228,7 +232,11 @@ namespace Keylight {
         InstanceId = instanceId,
         AppVersion = _config.AppVersion,
         SdkVersion = SdkInfo.Version,
-        Platform   = _config.Platform ?? Device.Platform
+        Platform   = _config.Platform ?? Device.Platform,
+        CpuCores   = Device.CpuCores,
+        OsVersion  = Device.OsVersionValue,
+        Arch       = Device.Arch,
+        Memory     = Device.Memory
       };
 
       ValidateResponse resp;